requirements.txt
# Web framework
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6  # For file upload handling

# Database
motor>=3.1.1  # MongoDB async driver
pymongo>=4.3.3

# Authentication and security
pyjwt>=2.6.0
bcrypt>=4.0.1
passlib>=1.7.4  # Password hashing
python-jose>=3.3.0  # JWT

# Data validation and serialization
pydantic>=1.10.7
email-validator>=2.0.0  # For email validation

# File operations
aiofiles>=23.1.0  # Async file operations

# Utilities
python-dotenv>=1.0.0  # Environment variables
typing-extensions>=4.5.0

# CORS
starlette>=0.26.1
